+++ botocore/httpsession.py
